# esr-mgr/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This page shows all the APIs available in the ESR server module.",
    "version" : "v1",
    "title" : "API Descriptions for External System Management",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Zi Li",
      "email" : "li.zi30@zte.com.cn"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath" : "/api/aai-esr-server/v1/",
...
  "paths" : {
...
    "/vims" : {
      "get" : {
        "tags" : [ " vim Management " ],
        "summary" : "get  all vim ",
        "description" : "",
        "operationId" : "queryVimList",
        "produces" : [ "application/json" ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ " vim Management " ],
        "summary" : "create a vim",
        "description" : "",
        "operationId" : "registerVims",
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain", "application/json" ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/vims/{cloudOwner}/{cloudRegionId}" : {
      "get" : {
        "tags" : [ " vim Management " ],
        "summary" : "get vim by id",
        "description" : "",
        "operationId" : "queryVimById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "cloudOwner",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "cloudRegionId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ " vim Management " ],
        "summary" : "update a vim",
        "description" : "",
        "operationId" : "updatevims",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "cloudOwner",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "cloudRegionId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ " vim Management " ],
        "summary" : "delete a vim",
        "description" : "",
        "operationId" : "delvim",
        "parameters" : [ {
          "name" : "cloudOwner",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "cloudRegionId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/vnfms" : {
      "get" : {
        "tags" : [ " vnfm Management " ],
        "summary" : "get  all vnfm ",
        "description" : "",
        "operationId" : "queryVnfmList",
        "produces" : [ "application/json" ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ " vnfm Management " ],
        "summary" : "create a vnfm",
        "description" : "",
        "operationId" : "registerVnfm",
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "vnfm",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/VnfmRegisterInfo"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/vnfms/{vnfmId}" : {
      "get" : {
        "tags" : [ " vnfm Management " ],
        "summary" : "get vnfm by id",
        "description" : "",
        "operationId" : "queryVnfmById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "vnfm id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ " vnfm Management " ],
        "summary" : "update a vnfm",
        "description" : "",
        "operationId" : "updateVnfm",
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "vnfm",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/VnfmRegisterInfo"
          }
        }, {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "vnfm id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ " vnfm Management " ],
        "summary" : "delete a vnfm",
        "description" : "",
        "operationId" : "delVnfm",
        "parameters" : [ {
          "name" : "vnfmId",
          "in" : "path",
          "description" : "vnfm id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "404" : {
            "description" : "microservice not found"
          },
          "415" : {
            "description" : "Unprocessable MicroServiceInfo Entity "
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "VnfmRegisterInfo" : {
      "type" : "object",
      "properties" : {
        "vnfmId" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "vimId" : {
          "type" : "string"
        },
        "vendor" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "certificateUrl" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "userName" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        }
      }
    }
  }
}